+++ R/hybridModel.R
@@ -6,6 +6,7 @@
 #' @import forecast
 #' @import stats
 #' @import graphics
+#' @import zoo
 #' @param y A numeric vector or time series.
 #' @param lambda
 #' Box-Cox transformation parameter.
@@ -162,16 +163,16 @@ hybridModel <- function(y, models = "aefnst",
     warning("auto.arima was not selected in the models argument, but a.args was passed. Ignoring a.args")
   }
   if(!is.null(e.args) && !is.element("e", expandedModels)){
-    warning("ets was not selected in the models argument, but e.args was passed. Ignoring a.args")
+    warning("ets was not selected in the models argument, but e.args was passed. Ignoring e.args")
   }
   if(!is.null(n.args) && !is.element("n", expandedModels)){
-    warning("nnetar was not selected in the models argument, but n.args was passed. Ignoring a.args")
+    warning("nnetar was not selected in the models argument, but n.args was passed. Ignoring n.args")
   }
   if(!is.null(s.args) && !is.element("s", expandedModels)){
-    warning("stlm was not selected in the models argument, but s.args was passed. Ignoring a.args")
+    warning("stlm was not selected in the models argument, but s.args was passed. Ignoring s.args")
   }
   if(!is.null(t.args) && !is.element("t", expandedModels)){
-    warning("tbats was not selected in the models argument, but t.args was passed. Ignoring a.args")
+    warning("tbats was not selected in the models argument, but t.args was passed. Ignoring t.args")
   }
 
   # Check for problems for specific models (e.g. long seasonality for ets and non-seasonal for stlm or nnetar)
@@ -538,19 +539,21 @@ accuracy.hybridModel <- function(f,
   return(forecast::accuracy(f$fitted, getResponse(f)))
 }
 
-#'Accuracy measures for cross-validated time series
+#' Accuracy measures for cross-validated time series
 #'
-#'Returns range of summary measures of the cross-validated forecast accuracy
-#'for \code{cvts} objects.
+#' Returns range of summary measures of the cross-validated forecast accuracy
+#' for \code{cvts} objects.
 #'
-#'@param f a \code{cvts} objected created by \code{\link{cvts}}.
-#'@param ... other arguments (ignored).
+#' @param f a \code{cvts} objected created by \code{\link{cvts}}.
+#' @param ... other arguments (ignored).
 #'
-#'@details
-#'Currently the method only implements \code{ME}, \code{RMSE}, and \code{MAE}. The accuracy measures
-#'\code{MPE}, \code{MAPE}, and \code{MASE} are not calculated. The accuracy is calculated for each
-#'forecast horizon up to \code{maxHorizon}
-#'@export
+#' @details
+#' Currently the method only implements \code{ME}, \code{RMSE}, and \code{MAE}. The accuracy measures
+#' \code{MPE}, \code{MAPE}, and \code{MASE} are not calculated. The accuracy is calculated for each
+#' forecast horizon up to \code{maxHorizon}
+#' @export
+#' @author David Shaub
+#' 
 accuracy.cvts <- function(f, ...){
   ME <- colMeans(f$residuals)
   RMSE <- apply(f$residuals, MARGIN = 2, FUN = function(x){sqrt(sum(x ^ 2)/ length(x))})